Renaming create webpack config for better clarity
